requirements.txt
discord~=1.0.1

python-dotenv~=0.15.0
